+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+#Change log 0.7.1
+*	Fix bug with RESTRequest factory in WebService class
+*	Result stream send by the server is now accessible within RESTRequest class by calling getResultStream()
+
 #Change log 0.7.0
 
 *	Request listeners now manage with RequestListeners class in order to avoid listener duplication